NSPX Comit 7
1. Log clean (xr_object_list.cpp, alife_update_manager.cpp, GameObject, xrServer_process_event_destroy, xrServer_process_event_ownership, Texture.cpp)
2. Add grass shadow fix by K.D. (DetailManager_VS.cpp)
3. Detail bump (uber_deffer.cpp)
4. Sun limit now 360 (was 180) in r2_R_sun.cpp

+++ xr_3da/xrImage_Resampler.cpp
@@ -263,7 +263,7 @@ void	imf_Process(u32* dstI, u32 dstW, u32 dstH, u32* srcI, u32 srcW, u32 srcH, E
 				contrib[i].n = 0;
 				contrib[i].p = (CONTRIB*)xr_malloc((int)(width * 2 + 1) * sizeof(CONTRIB));
 				ZeroMemory(contrib[i].p, (int)(width * 2 + 1) * sizeof(CONTRIB));
-				center = float(i) / xscale;
+				center = double(i) / xscale;
 				left = ceil(center - width);
 				right = floor(center + width);
 				for (j = int(left); j <= int(right); ++j)
@@ -296,7 +296,7 @@ void	imf_Process(u32* dstI, u32 dstW, u32 dstH, u32* srcI, u32 srcW, u32 srcH, E
 				contrib[i].n = 0;
 				contrib[i].p = (CONTRIB*)xr_malloc((int)(fwidth * 2 + 1) * sizeof(CONTRIB));
 				ZeroMemory(contrib[i].p, (int)(fwidth * 2 + 1) * sizeof(CONTRIB));
-				center = float(i) / xscale;
+				center = double(i) / xscale;
 				left = ceil(center - fwidth);
 				right = floor(center + fwidth);
 				for (j = int(left); j <= int(right); ++j)
@@ -375,7 +375,7 @@ void	imf_Process(u32* dstI, u32 dstW, u32 dstH, u32* srcI, u32 srcW, u32 srcH, E
 				contrib[i].n = 0;
 				contrib[i].p = (CONTRIB*)xr_malloc((int)(width * 2 + 1) * sizeof(CONTRIB));
 				ZeroMemory(contrib[i].p, (int)(width * 2 + 1) * sizeof(CONTRIB));
-				center = (float)i / yscale;
+				center = (double)i / yscale;
 				left = ceil(center - width);
 				right = floor(center + width);
 				for (j = int(left); j <= int(right); ++j)
@@ -406,7 +406,7 @@ void	imf_Process(u32* dstI, u32 dstW, u32 dstH, u32* srcI, u32 srcW, u32 srcH, E
 				contrib[i].n = 0;
 				contrib[i].p = (CONTRIB*)xr_malloc((int)(fwidth * 2 + 1) * sizeof(CONTRIB));
 				ZeroMemory(contrib[i].p, (int)(fwidth * 2 + 1) * sizeof(CONTRIB));
-				center = (float)i / yscale;
+				center = (double)i / yscale;
 				left = ceil(center - fwidth);
 				right = floor(center + fwidth);
 				for (j = int(left); j <= int(right); ++j) {
